requirements.txt
# Web server
flask>=3.0,<4.0
waitress>=3.0,<4.0

# Google APIs
google-api-python-client>=2.100,<3.0
google-auth>=2.20,<3.0
google-auth-oauthlib>=1.2,<2.0
google-auth-httplib2>=0.2,<1.0

# AI Providers
openai>=1.50,<2.0
boto3>=1.35,<2.0

# Data validation & config
pydantic>=2.5,<3.0
pydantic-settings>=2.1,<3.0
python-dotenv>=1.0,<2.0

# Resilience
tenacity>=9.0,<10.0

# HTML parsing
beautifulsoup4>=4.12,<5.0

# Testing
pytest>=8.0,<9.0
pytest-mock>=3.12,<4.0
responses>=0.25,<1.0
